requirements.txt
easyocr
easygui
matplotlib
pillow
google.generativeai
